server/pyproject.toml
[tool.poetry]
package-mode = false
[project]
name = "server"
version = "0.1.0"
description = "FastAPI backend for Avkash HRMS with MongoDB Atlas integration"
authors = [
    {name = "username",email = "emailid@domain.com"}
]
requires-python = ">=3.10,<=3.13.3"
dependencies = [
    "uvicorn[standard]>=0.34.2,<0.35.0",
    "pymongo[srv] (>=4.13.0,<5.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "requests (>=2.32.3,<3.0.0)",
    "sentence-transformers (>=4.1.0,<5.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "pdfplumber (>=0.11.6,<0.12.0)",
    "python-docx (>=1.1.2,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "websockets>=12.0,<13.0",
    "apscheduler (>=3.11.0,<4.0.0)",
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "pymupdf (>=1.26.1,<2.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "joblib (>=1.5.1,<2.0.0)",
    "scikit-image (>=0.25.2,<0.26.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
pytest = "^8.3.5"
numpy = "==2.2.6"
scipy = "==1.15.3"

